requirements.txt
WTForms==2.3.3
Flask_Login==0.5.0
alembic==1.4.3
Flask_Bootstrap==3.3.7.1
Werkzeug==1.0.1
flask_nav==0.6
Flask==1.1.2
Flask_WTF==0.14.3
SQLAlchemy==1.3.20
Flask_Migrate==2.5.3
Flask_Minify==0.28
Flask_SQLAlchemy==2.4.4
flask_talisman==0.7.0
Flask_Obscure==1.0
python-dotenv==0.15.0
email-validator==1.1.1
# gevent server
gevent==20.9.0